2022 complete up to day 11
The solution for day 11 is sketchy.
As is, it works for part 2, but not for part 1.

At first, the part 1 code simply represented each stress level as an
integer. This works for part 1, but in part 2 the numbers become too
large, large enough that if we use Integer we run out of memory.

So I found a more intelligent representation for the stress level and
was able to come up with a solution for part 2. The thing is, after
refactoring to represent the stress level differently, I was no longer
able to calculate the solution to part 1.

I'm unsure why that is. I arrived at the new representation by looking
at all the things I needed to be able to do with a stress level. I
identified five (arithmetic) operations. I implemented these
operations for the new representation. The solution calculated for
part 2 is correct! But not the solution calculated for part 1 is
slightly too high.

I have a hunch. I believe it's the rounding down behaviour that is
expected when we divide the stress level by three. I ran a bunch of
tests and the behaviour was observed. Maybe it takes a very special
situation for it to not round down.